+++ InputfieldHelper.module.php
@@ -6,6 +6,32 @@
  * @author			: İskender TOTOĞLU, @ukyo (community), @trk (Github)
  * @website			: https://www.altivebir.com
  * @projectWebsite	: https://github.com/trk/InputfieldHelper
+ *
+ * @method $this form_class($value = "")
+ * @method $this list_class($value = "")
+ * @method $this list_clearfix_class($value = "")
+ * @method $this item_class($value = "")
+ * @method $this item_label_class($value = "")
+ * @method $this item_content_class($value = "")
+ * @method $this item_required_class($value = "")
+ * @method $this item_error_class($value = "")
+ * @method $this item_collapsed_class($value = "")
+ * @method $this item_column_width_class($value = "")
+ * @method $this item_column_width_first_class($value = "")
+ * @method $this item_show_if_class($value = "")
+ * @method $this item_required_if_class($value = "")
+ *
+ * @method $this list_markup($value = "")
+ * @method $this item_markup($value = "")
+ * @method $this item_label_markup($value = "")
+ * @method $this item_label_hidden_markup($value = "")
+ * @method $this item_content_markup($value = "")
+ * @method $this item_error_markup($value = "")
+ * @method $this item_description_markup($value = "")
+ * @method $this item_head_markup($value = "")
+ * @method $this item_notes_markup($value = "")
+ * @method $this item_icon_markup($value = "")
+ * @method $this item_toggle_markup($value = "")
  */
 class InputfieldHelper extends ModuleConfig implements Module {
 
@@ -15,6 +41,109 @@ class InputfieldHelper extends ModuleConfig implements Module {
      * @var null
      */
     public $wrapper = null;
+
+    /**
+     * Clear InputfieldWrapper Markup
+     *
+     * @var bool
+     */
+    public $resetMarkup = false;
+
+    /**
+     * Markups
+     *
+     * @var string
+     */
+    public $markup = array();
+
+    /**
+     * Clean Markup
+     *
+     * @var array
+     */
+    protected $cleanMarkup = array(
+        "list" => "{out}",
+        "item" => "{out}",
+        "item_label" => "{out}",
+        "item_label_hidden" => "{out}",
+        "item_content" => "{out}",
+        "item_error" => "{out}",
+        "item_description" => "{out}",
+        "item_head" => "{out}",
+        "item_notes" => "{out}",
+        "item_icon" => "",
+        "item_toggle" => ""
+    );
+
+    /**
+     * Clear InputfieldWrapper classes
+     *
+     * @var bool
+     */
+    public $resetClasses = false;
+
+    /**
+     * Clean Classes
+     *
+     * @var array
+     */
+    protected $cleanClasses = array(
+        "form" => "",
+        "list" => "",
+        "list_clearfix" => "",
+        "item" => "{class} {name}",
+        "item_label" => "",
+        "item_content" => "",
+        "item_required" => "",
+        "item_error" => "",
+        "item_collapsed" => "",
+        "item_column_width" => "",
+        "item_column_width_first" => "",
+        "item_show_if" => "",
+        "item_required_if" => ""
+        // ALSO:
+        // InputfieldAnything => array( any of the properties above to override on a per-Inputifeld basis)
+    );
+
+    /**
+     * Classes
+     *
+     * @var string
+     */
+    public $classes = array();
+
+    /**
+     * Inputfield Variations
+     *
+     * @var array
+     */
+    protected $inputfield = array(
+        "markup" => array(),
+        "classes" => array(),
+        "append" => "",
+        "prepend" => ""
+    );
+
+    /**
+     * Front-end Framework
+     *
+     * @var string $this->frameworks["uikit3"]
+     */
+    public $framework = "uikit3";
+
+    /**
+     * Breakpoint for Front-end framework
+     *
+     * @var string
+     */
+    public $frameworkBreakpoint = "@m";
+
+    /**
+     * Front-end Frameworks
+     *
+     * @var array
+     */
+    protected $frameworks = array("uikit2", "uikit3");
 
     /**
      * Inputfield Name Prefix
@@ -43,6 +172,13 @@ class InputfieldHelper extends ModuleConfig implements Module {
      * @var string
      */
     public $id_suffix = "";
+
+    /**
+     * If not null apply collapsed value to all inputfields
+     *
+     * @var null
+     */
+    public $collapsed = null;
 
     /**
      * Inputfields Values
@@ -76,8 +212,8 @@ class InputfieldHelper extends ModuleConfig implements Module {
             // "permission" => "permission-name",
             // "permissions" => array()
             "icon" => "steam",
-            // "singular" => false,
-            // "autoload" => false
+            "singular" => true,
+            "autoload" => true
         );
     }
 
@@ -90,6 +226,214 @@ class InputfieldHelper extends ModuleConfig implements Module {
      */
     public function __construct() {
 
+    }
+
+    /**
+     * Set markup or classes with magic function call
+     *
+     * @param string $name
+     * @param array $args
+     * @return $this
+     */
+    public function __call($name = "", $args = array()) {
+        $value = $this->element(0, $args, "");
+        if(strpos($name, "_markup") !== FALSE) {
+            $this->setMarkup(str_replace("_markup", "", $name), $value);
+        }
+        if(strpos($name, "_class") !== FALSE) {
+            $this->setClass(str_replace("_class", "", $name), $value);
+        }
+        return $this;
+    }
+
+    // ------------------------------------------------------------------------
+
+    /**
+     * Set element markups
+     *
+     * @param string $key
+     * @param string $value
+     */
+    public function setMarkup($key = "", $value = "") {
+        $this->markup[$key] = $value;
+    }
+
+    // ------------------------------------------------------------------------
+
+    /**
+     * Set element classes
+     *
+     * @param string $key
+     * @param string $value
+     */
+    public function setClass($key = "", $value = "") {
+        $this->classes[$key] = $value;
+    }
+
+    // ------------------------------------------------------------------------
+
+    public function ___prepareInputfields() {
+        $this->addHookBefore('InputfieldWrapper::render', function($event) {
+            $inputfieldWrapper = $event->object;
+
+            // Clean markup and classes
+            if($this->resetMarkup) $inputfieldWrapper->setMarkup($this->cleanMarkup);
+            if($this->resetClasses) $inputfieldWrapper->setClasses($this->cleanClasses);
+
+            // Get markup and classes
+            $markupWrapper = $inputfieldWrapper->getMarkup();
+            $classesWrapper = $inputfieldWrapper->getClasses();
+
+            // Set markup and classes
+            if(is_array($this->markup) && count($this->markup)) {
+                $markupWrapper = array_merge($markupWrapper, $this->markup);
+                $inputfieldWrapper->setMarkup($markupWrapper);
+            }
+            if(is_array($this->classes) && count($this->classes)) {
+                $classesWrapper = array_merge($classesWrapper, $this->classes);
+                $inputfieldWrapper->setClasses($classesWrapper);
+            }
+
+            foreach ($inputfieldWrapper->children as $inputfield) {
+                $name = $inputfield->name;
+                // Get inputfield
+                $inputfieldArray = $this->element($name, $this->inputfieldsArray, array());
+
+                $append = $this->element("append", $inputfieldArray, "");
+                $prepend = $this->element("prepend", $inputfieldArray, "");
+                $markup = $this->element("markup", $inputfieldArray, array());
+                $classes = $this->element("classes", $inputfieldArray, array());
+
+                // Field name based markup
+                if(count($markup)) {
+                    $inputfieldMarkup = array_merge($markupWrapper, $markup);
+                    $inputfieldWrapper->setMarkup($inputfieldMarkup);
+                } else {
+                    $inputfieldWrapper->setMarkup($markupWrapper);
+                }
+                // Field name based classes
+                if(count($classes)) {
+                    $inputfieldClasses = array_merge($classesWrapper, $classes);
+                    $inputfieldWrapper->setClasses($inputfieldClasses);
+                } else {
+                    $inputfieldWrapper->setClasses($classesWrapper);
+                }
+
+                if(in_array($this->framework, $this->frameworks)) {
+                    if($this->framework == "uikit2" || $this->framework == "uikit3") {
+                        if($inputfield instanceof InputfieldTextarea) {
+                            $inputfield->addClass('uk-textarea');
+                        } else if($inputfield instanceof InputfieldSelect) {
+                            $inputfield->addClass('uk-select');
+                        } else if($inputfield instanceof InputfieldPassword) {
+                            $inputfield->addClass('uk-input uk-form-width-medium');
+                        } else if($inputfield instanceof InputfieldText || $inputfield instanceof InputfieldInteger) {
+                            $inputfield->addClass('uk-input');
+                        } else if($inputfield instanceof InputfieldCheckboxes || $inputfield instanceof InputfieldCheckbox) {
+                            $inputfield->addClass('uk-checkbox');
+                            $inputfield->addClass('uk-form-controls-text', 'contentClass');
+                        } else if($inputfield instanceof InputfieldRadios) {
+                            $inputfield->addClass('uk-radio');
+                            $inputfield->addClass('uk-form-controls-text', 'contentClass');
+                        } else if($inputfield instanceof InputfieldSubmit || $inputfield instanceof InputfieldButton) {
+                            $inputfield->addClass('uk-button uk-button-primary');
+                        } else if($inputfield instanceof InputfieldHidden) {
+                            $inputfield->wrapAttr('class', 'uk-hidden');
+                        }
+                    }
+
+                    if($this->framework == "uikit2") {
+                        $inputfield->wrapAttr('class', $this->frameworkUikit2($inputfield->columnWidth));
+                    } elseif($this->framework == "uikit3") {
+                        $inputfield->wrapAttr('class', $this->frameworkUikit3($inputfield->columnWidth));
+                    }
+                }
+
+                if($append || $prepend) {
+                    $inputfield->addHookAfter('render', function($inputfieldEvent) use($append, $prepend) {
+                        $inputfieldEvent->replace = true;
+                        $inputfieldEvent->return = $prepend . $inputfieldEvent->return . $append;
+                    });
+                }
+            }
+        });
+    }
+
+    // ------------------------------------------------------------------------
+
+    /**
+     * Uikit 2 Front-end framework for modifier
+     *
+     * @param string $columnWidth
+     * @return string
+     */
+    protected function frameworkUikit2($columnWidth="") {
+        $width = "";
+        $class = "uk-width-1-1";
+        if($columnWidth != "") {
+            if($columnWidth > 0 && $columnWidth < 11) {
+                $width = 1;
+            } elseif ($columnWidth > 10 && $columnWidth < 21) {
+                $width = 2;
+            } elseif ($columnWidth > 20 && $columnWidth < 31) {
+                $width = 3;
+            } elseif ($columnWidth > 30 && $columnWidth < 41) {
+                $width = 4;
+            } elseif ($columnWidth > 40 && $columnWidth < 51) {
+                $width = 5;
+            } elseif ($columnWidth > 50 && $columnWidth < 61) {
+                $width = 6;
+            } elseif ($columnWidth > 60 && $columnWidth < 71) {
+                $width = 7;
+            } elseif ($columnWidth > 70 && $columnWidth < 81) {
+                $width = 7;
+            } elseif ($columnWidth > 80 && $columnWidth < 91) {
+                $width = 9;
+            } elseif ($columnWidth > 90) {
+                $width = 10;
+            }
+            if($width == 10) $class = "uk-width-1-1";
+            else $class = "uk-width-{$this->frameworkBreakpoint}{$width}-10";
+        }
+        return $class;
+    }
+
+    // ------------------------------------------------------------------------
+
+    /**
+     * Uikit 3 inputfield modifier for ProcessWire Inputfields
+     *
+     * @param string $columnWidth
+     * @return string
+     */
+    protected function frameworkUikit3($columnWidth = "") {
+        $ukWidthClass = '1-1';
+        // determine column width class
+        // uk class => width %
+        $ukGridWidths = array(
+            "100%" => "1-1",
+            "83%" => "5-6",
+            "80%" => "4-5",
+            "60%" => "3-5",
+            "50%" => "1-2",
+            "33%" => "1-3",
+            "25%" => "1-4",
+            "20%" => "1-5",
+            "16%" => "1-6"
+        );
+
+        if($columnWidth && $columnWidth < 100) {
+            if($columnWidth < 16) $columnWidth = 16;
+            foreach($ukGridWidths as $pct => $uk) {
+                $pct = (int) $pct;
+                if($columnWidth >= $pct) {
+                    $ukWidthClass = $uk;
+                    break;
+                }
+            }
+        }
+
+        return "uk-width-" . $ukWidthClass . $this->frameworkBreakpoint;
     }
 
     // ------------------------------------------------------------------------
@@ -109,6 +453,7 @@ class InputfieldHelper extends ModuleConfig implements Module {
      * @throws WireException
      */
     public function getInputfields() {
+        $this->___prepareInputfields();
         foreach($this->getDefaults() as $key => $value) {
             $this->set($key, $value);
         }
@@ -131,7 +476,7 @@ class InputfieldHelper extends ModuleConfig implements Module {
      * @param array $inputfields
      * @return array
      */
-    public function prepareInputfields($inputfields = array()) {
+    protected function prepareInputfields($inputfields = array()) {
         if(count($inputfields)) {
             foreach ($inputfields as $key => $inputfield) {
                 $children = $this->element("children", $inputfield, array());
@@ -143,6 +488,7 @@ class InputfieldHelper extends ModuleConfig implements Module {
                 $inputfields[$key]["id"] = $this->id_prefix . $id . $this->id_suffix;
                 $inputfields[$key]["name"] = $this->name_prefix . $name . $this->name_suffix;
                 $inputfields[$key]["value"] = $value;
+                if(!is_null($this->collapsed)) $inputfields[$key]["collapsed"] = $this->collapsed;
 
                 if($showIf = $this->element('showIf', $inputfield, '')) {
                     $inputfields[$key]["showIf"] = $this->showIf($showIf);
@@ -192,5 +538,16 @@ class InputfieldHelper extends ModuleConfig implements Module {
      */
     public function element($item, array $array, $default = NULL) {
         return array_key_exists($item, $array) && $array[$item] ? $array[$item] : $default;
+    }
+}
+
+if(!function_exists('iHelper')) {
+    /**
+     * Shortcut function for InputfieldHelper class
+     *
+     * @return InputfieldHelper
+     */
+    function iHelper() {
+        return new InputfieldHelper();
     }
 }